requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.20
python-dotenv>=1.0.0
alembic==1.12.1
psycopg2-binary>=2.9.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.6
openai==1.3.5
cors==1.0.1
pydantic>=2.3.0
pydantic-settings>=2.0.3
email-validator==2.1.0.post1 